backend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==3.2.16
django-cors-headers==4.4.0
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
flake8==7.1.0
idna==3.7
isort==5.13.2
mccabe==0.7.0
oauthlib==3.2.2
pillow==10.4.0
psycopg2-binary==2.9.3
pycodestyle==2.12.0
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
